updated the integer type of balance from integer to float

+++ core/database/migrations/2021_11_10_102705_create_parent_due_balances_table.php
@@ -17,7 +17,7 @@ class CreateParentDueBalancesTable extends Migration
             $table->uuid('uuid')->primary();
             $table->string('parent_name',60);
             $table->string('national_id')->unique()->index();
-            $table->integer('balance');//will mutate it by multiply it by 1000 and return divide it by 1000 when accessing
+            $table->float('balance');//will mutate it by multiply it by 1000 and return divide it by 1000 when accessing
             $table->string('email')->unique()->index();
             $table->timestamps();
         });
